inline ChartBeat code inside output block
